+++ src/hoarder/archives/__init__.py
@@ -1,9 +1,9 @@
+from .abstract_rar_archive import AbstractRarArchive, RarArchiveError
 from .hash_archive import Algo, FileEntry, HashArchive
 from .hash_archive_repository import HashArchiveRepository
 from .hash_name_archive import HashNameArchive
-from .rar_archive import AbstractRarArchive, RarArchiveError
+from .rar_archive import RarArchive
 from .rar_archive_7z import Rar7zArchive
-from .rar_archive_default import RarArchive
 from .rar_archive_rarfile import RarfileRarArchive
 from .rar_path import RarScheme
 from .sfv_archive import SfvArchive
@@ -21,11 +21,11 @@ __all__ = [
     "RarfileRarArchive",
     "RarScheme",
     "SfvArchive",
+    "abstract_rar_archive",
     "hash_archive",
     "hash_archive_repository",
     "rar_archive",
     "rar_archive_7z",
-    "rar_archive_default",
     "rar_archive_rarfile",
     "sfv_archive",
 ]
