+++ src/hoarder/archives/rar_archive_rarfile.py
@@ -7,8 +7,8 @@ import zlib
 
 import rarfile
 
+from .abstract_rar_archive import AbstractRarArchive, RarArchiveError
 from .hash_archive import Algo, FileEntry
-from .rar_archive import AbstractRarArchive, RarArchiveError
 from .rar_path import locate_main_volume
 
 try:
